backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
clerk-python
svix
openai
langchain
langchain-openai 